Refactor: Moved shared data cleaning functions to new file called  preprocessing_utils.py
