+++ tennis_racquet_analysis/dataset.py
@@ -2,34 +2,15 @@ from pathlib import Path
 from loguru import logger
 from tqdm import tqdm
 import typer
-import pandas as pd
 from tennis_racquet_analysis.config import RAW_DATA_DIR, INTERIM_DATA_DIR
+from tennis_racquet_analysis.preprocessing_utils import (
+    load_data,
+    drop_column,
+    rename_column,
+    squared,
+)
 
 app = typer.Typer()
-
-
-def load_data(input_path: Path) -> pd.DataFrame:
-    logger.info(f"Looking for file at: {input_path}")
-    if input_path.exists():
-        df = pd.read_csv(input_path)
-        logger.info("Data loaded successfully!")
-        return df
-    else:
-        raise FileNotFoundError(f"File not found. Please check your path: {input_path}")
-
-
-def drop_column(dataframe, column):
-    return dataframe.drop(columns=[column])
-
-
-def rename_column(dataframe, column):
-    new_column = column.replace(".", "")
-    return dataframe.rename(columns={column: new_column})
-
-
-def squared(dataframe, column):
-    dataframe[f"{column}_sq"] = dataframe[column] ** 2
-    return dataframe
 
 
 @app.command()
